Added some more files

+++ minishell.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   minishell.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ekulichk <ekulichk@student.42.fr>          +#+  +:+       +#+        */
+/*   By: mtravez <mtravez@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/01 14:02:34 by mtravez           #+#    #+#             */
-/*   Updated: 2023/04/30 10:34:59 by ekulichk         ###   ########.fr       */
+/*   Updated: 2023/05/11 13:02:37 by mtravez          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ void	print_tokens(t_lexer *lexer)
 	token = lexer->token;
 	while (token)
 	{
-		ft_printf("%s, [%i]\n", token->content, token->t_type);
+		ft_printf("(%s), [%i]\n", token->content, token->t_type);
 		token = token->next_token;
 	}
 }
@@ -50,7 +50,19 @@ t_msvars	*init_ms(void)
 	return (ms);
 }
 
-int	main(void)
+char *get_smth(char *str, char **env)
+{
+	int	i = 0;
+	while (env && env[i])
+	{
+		if (ft_strncmp(str, env[i], ft_strlen(str)) == 0)
+			return (env[i]);
+		i++;
+	}
+	return (NULL);
+}
+
+int	main(int argc, char **argv, char **env)
 {
 	t_msvars	*mini_shell;
 	char		*lineptr;
@@ -59,16 +71,18 @@ int	main(void)
 
 	mini_shell = init_ms();
 	lineptr = readline(PROMPT);
+	if (!argc || !argv || !env)
+		return (0);
 	while (lineptr)
 	{
 		if (ft_strlen(lineptr) > 0)
 			add_history(lineptr);
-		lexer = get_tokens(lineptr, ft_strlen(lineptr) + 1);
-		// print_tokens(lexer);
-		ft_printf("%s\n", lineptr);
+		lexer = get_tokens(lineptr, ft_strlen(lineptr));
+		print_tokens(lexer);
+		// ft_printf("%s\n", lineptr);
 		free(lineptr);
 		root = parse_to_tree(lexer, get_last(lexer->token));
-		print_tree(root, 0);
+		// print_tree(root, 0);
 		destroy_lexer(lexer);
 		free_tree(root);
 		lineptr = NULL;
